+++ src/main/java/com/tool/RecruitXpert/Repository/UserRepository.java
@@ -4,7 +4,11 @@ import com.tool.RecruitXpert.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User,Integer> {
     User findByEmail(String email);
+
+//    List<User> findById(int userId);
 }
